tests/requirements-test.txt
# Testing Requirements for REE AI Platform

# Core Testing Framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-timeout==2.2.0
pytest-xdist==3.5.0  # Parallel test execution

# HTTP Testing
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# AI/ML Testing
numpy==1.26.2
scikit-learn==1.3.2  # For metrics and evaluation

# Mock and Fixtures
pytest-mock==3.12.0
faker==20.1.0
factory-boy==3.3.0

# Performance Testing
locust==2.20.0
pytest-benchmark==4.0.0

# Test Utilities
python-dotenv==1.0.0
pydantic==2.5.2
tenacity==8.2.3  # Retry logic for flaky tests

# Reporting
pytest-html==4.1.1
pytest-json-report==1.5.0
allure-pytest==2.13.2

# Code Quality
black==23.12.1
flake8==7.0.0
mypy==1.7.1
